Fix some ngtcp2 usage and conn issue detection
Clients now get an ERR_PROTO packet when something goes wrong and the
server is unhappy.

ngtcp2_accept changed between our last update of it and current, this
fixes its usage.

With a minor change to the C API, liblokinet userland will now be able
to receive notification when a connection is (un)successfully opened,
and when it closes.  TODO: supply a reason for the close, if we decide
it should.