just the headers for the internals refactor.
llarp/layers/platform/*.hpp

was the "front end" of the previous kitchen sink.

this component bridges the userland platforms, be the full VPN mode,
embedded mode or any other kind of future mode of operation we may
support. this component is meant to slowly consilidate the location of
platform specific quarks, like ones for proactor vs reactor event
loops. the vpn platform code will exist independantly from this layer
but this will be where any kind of idempotent queue flushing when we
get something from or destinted the "platform" (e.g. the vpn
interface, or some kind of internal for embedded lokinet).

llarp/layers/flow/*.hpp

was the "back end" of the previous kitchen sink.

we provide all headers for the logic of .loki and .snode traffic
codepaths for snodes and clients. these headers support .loki and
.snode in a unified api, any additional kinds of remote endpoint types
like a .exit whatever else would be easy to add with this.

defines hooks for any authentication between us and remote flows,
regardless of if we are initiator or recipient.

defines ons name to flow address lookup interface, as well as ons name
caching interface. both of which adds a slot to define future
configurable knobs how strict of a consensus we want for a ons lookup
and result caching lifetime and lookup result
invalidation (e.g. allows us to purge things when we add consensus
state hints into lokinet clients).

llarp/layers/route/*.hpp

the routing layer which used to blur into the "back end" old kitchen
sink (now called the flow layer). currently mostly baren and exists as
a placeholder stub until later.

llarp/layers/onion/*.hpp

stub for onion layer, basically for llarp/path/* as a mechanism for
removing llarp::path::PathSet type forever. this has yet to happen.

llarp/context.hpp

this was include/llarp.hpp, but now it is here and no longer a public
header. we really dont have a nice way to expose a C++ api just a C
api. going forward a public C++ api does not seem to be a worthwhile
efffort for abi stability reasons and how crap our current headers are.

will remove include/llarp.hpp later as so much includes it.

(maybe C++ modules will help later... maybe...)