fix bad_variant_access resulting from ConvoTag->AddressVariant change
We changed the quic code to refer to its remote endpoint using an Address Variant
and this resulted in a crash when sending a connection close packet.  This code was
incorrectly supplying an uninitialized value for ConvoTag before but this was silently
ignored; after switching to a variant it became a crash.