fix race when initializing context in lokinet main.
when we configure but fail to start up, we can hit a case when we
have a context that is configured but not setup or running. when we
try to sigterm/sigint we get stuck and cannot exit as the
llarp::Context shared_ptr global is nullptr but dangles elsewhere.

we make it so that we fufill a promose if we fully configure and setup
successfully right before running the event loop main for it. this
makes it so we bail if setup was not completed and exit our process.