Redo wire protocol race condition fix.
In commit a76acd49560f9845c2ad9a839364f2834afc7756 we attempted to fix
issues on service nodes that related to spamming service nodes with
traffc, but this caused a regression where clients could not connect
to the network.

However, we also did something idiotic with the protocol handshake.
The handshake logic was modified and caused client connections to
fail. This was due to a change in how we decided what a timed out
session is. We redo the close bug and revert the change to the
protocol handshake.

(cherry picked from commit b0b6e2b198ba32e8c5641e99494d2874b2243b8f)