Prefix cmake settings with LOKINET_
There are significant conflicts with parent project variables here when
trying to use lokinet as a submodule.  This renames most of the options
to be `LOKINET_...`, e.g. `-DLOKINET_TESTS=ON`.  A few aren't renamed
because they are common option names (e.g. BUILD_SHARED_LIBS) that
aren't likely to differ across parent project and submodule.

Where the current option started with `USE_` or `WITH_` or `BUILD_` I
removed that (so that `WITH_TESTS` becomes `LOKINET_TESTS`).  The
`WITH_PEERSTATS` option was sometimes misreferenced as
`WITH_PEERSTATS_BACKEND`, but now both are `LOKINET_PEERSTATS`.

Also removes some options that do nothing (or no longer do anything):

- `USE_NETNS` -- this was a not-finished idea way back when to use Linux
  network namespaces.
- `TESTNET` -- this was to run a local testnet on a single machine, but
  no longer has any effect (we can always resurrect it if we resurrect
  that capability).