Don't force embedded liblokinet to be a shared lib
We most likely don't want a .so for libsession-util, so don't force one,
but rather just let it do whatever BUILD_SHARED_LIBS instructs.

Rename the target to `lokinet-embedded` because `lokinet-shared` makes
no sense for a target that might or might not be shared.

Also adds a `lokinet::embedded` alias for a parent project to use.