Category logging & don't-touch-logging option
This adds category logging configuration support to lokinet so that you
can specify things like "*=warn, quic=debug" to more finely control log
output.

It also adds an ability for the logging type to be set to nullopt so
that lokinet can be initialized without nuking all existing log sinks,
such as from libsession, and makes this the default for an embedded
config.