WIP tunneled TCP
I believe all of the *ideas* needed for TCP tunneling are present, but
at the moment it segfaults, so I can't test that and say for sure.

Also adds an appended byte to session datagrams to indicate what that
datagram is.  Happy to change the format, but appending is useful so we
can just truncate and then move the buffer if we want.