Not all packets are IPv4 or IPv6
IPPacket parsing was assuming that any packet that isn't IPv4 must be
IPv6, but that isn't true.

This adds proper failure for other types of packets.

It also reduces the not-supported warning on IPv6 packets because we
actually get a fair number of them on Linux to the kernel's link-local
address that just fill up logs needlessly with warnings about them.